envs/verify/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.4.2,<3.0.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
httpx==0.26.0

